Reorganis the package structure to use subdirectories cleanly
