+++ pycufsm/pre/geometry.py
@@ -3,6 +3,8 @@ from typing import Literal, Optional
 
 import numpy as np
 
+# Note that this file is entirely new to pyCUFSM, and similar functions will NOT be found in the MATLAB version of CUFSM.
+
 ANGLE_TOLERANCE = np.radians(5)  # Tolerance for detecting a corner in radians
