+++ pycufsm/solve/analysis.py
