+++ pycufsm/solve/analysis_p.py
