+++ pycufsm/solve/cfsm.py
